Move work queue to shared module.
