--- shared/test.go
@@ -1,7 +1,0 @@
-package shared
-
-import "fmt"
-
-func Test() {
-	fmt.Println("Hello from the shared module.")
-}
